LocalDB: rename files: data_cleanup to vacuum
