+++ ydb/core/tablet_flat/ut/ya.make
@@ -38,7 +38,6 @@ SRCS(
     ut_comp_gen.cpp
     ut_compaction.cpp
     ut_compaction_multi.cpp
-    ut_data_cleanup.cpp
     ut_datetime.cpp
     ut_decimal.cpp    
     ut_charge.cpp
@@ -56,6 +55,7 @@ SRCS(
     ut_shared_sausagecache_actor.cpp
     ut_slice.cpp
     ut_slice_loader.cpp
+    ut_vacuum.cpp
     ut_versions.cpp
 )
